+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Expire date shown for product in list table 
+
 ### 1.2.0: December 28th, 2020
 * Easy Updates Manager expiry notice
 * Remove Woo Store Vacation upsell notice